chunked: add fallback for openat2
add a fallback mechanism when openat2 is not supported by the
underlying kernel.

If a call to openat2 fails with ENOSYS, then fallback to a user space
lookup.  Generally the user space lookup is not safe, since symlinks
lookups are vulnerable to TOCTOU attacks, but in this case where the
rootfs is being created, there are no other processes modifying it
thus such lookups can be considered safe.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>