libimage: fix reference filters
It turns out that FamiliarMatch is only useful for matching Docker Hub
but we should not limit it to that and match values against registry.

For instance, FamiliarMatch is *not* able to match a FQN reference
against a Docker Hub image.  I am convinced that we should *not* behave
as Docker does in this case.

This brings us back to the behavior prior to commit c15939746c59 but
with a fixed matching algorithm.  The specified value will now be
matched against 1) the FQN 2) without domain 3) without domain and path.
If specified also a second time without digest/tag.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>