As a fail-safe, terminate recreateSymlinks after 10 iterations
The code can only do 3 kinds of corrective actions per layer
(we probably could order them so that it doesn't need to loop
at all); so, after clearly more iterations than that, assume
the filesystem is somehow broken (e.g. deleting a file "succeeds"
but the old value is still around), or there is a bug in the
body of the loop, and report the error instead of looping
forever.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>