Log all kinds of failures to delete a layer
Currently, if the attempts to recover from a failure
themselves fail, we don't record that at all.

That makes diagnosing the situation, or hypothetically
detecting that the cleanup could never work, much
harder.

So, log the errors.

Alternatively, we could include those failures as extra
text in the returned error; that's less likely to be lost
(e.g. a Go caller would have the extra text available, without
setting up extra infrastructure to capture logs), but possibly
harder to follow.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>