Remove keyring support
The package is not used, noone is currently working on wiring it up,
and it started failing in CI (apparently because it assumes the user's
keyring is empty when tests start, which is now not reliably the case).

Instead of fixing the tests, just remove it; we can always restore
the code from Git history.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>