overlay: Account for the trailing '\0' in mount() string.
There's some code here that checks if we've overflown
the amount of data that mount() can take (it's bounded
by the RAM page size, I guess because it's a syscall?
I don't really understand but I'm okay with that). But
it's not quite right: if you happened to land on the page
size exactly it would crash.

Fixes https://github.com/containers/podman/issues/13432

Signed-off-by: Nick Guenther <nick.guenther@polymtl.ca>