ci/gha: add lint job
This is currently similar to what make validate does in .cirrus.yml,
but let's have it as a separate job since it it faster.

Let's keep the same job in Cirrus CI for now.

Note we do not ping Go version, but rather depend on action/setup-go to
know which one is latest.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>