Introduce manifestConversionPlan
To make determineManifestConversion easier to test,
we'll want to separate it from imageCopier and make it a pure function.

To do that, start with introducing a struct for its return value;
that way we won't need three unnamed return values, and three
variables with long names at the call site.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>