Record an *image.SourcedImage in imageCopier.src
This is surprisingly somewhat convoluted.

We've had to do a bunch of work to make determineManifestConversion
not depend on a complete (and mostly valid!) image, to be able
to make this change.

Then, so that we don't lose test coverage (and to actually benefit
from the new private type), use SourcedImage.ManifestMIMEType
in imageCopier.determineManifestConversion.  That allows us to get
rid of the last imageCopier.determineManifestConversion call
in tests.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>