overlay: create linkDir with mask 0755
on newer kernels we use idmapped mounts for the overlayfs.

When running in a user namespace that doesn't include the root user
from the host, the linDir cannot be accessed since CAP_DAC_OVERRIDE
doesn't work with IDs that are not part of the mapping.

To avoid this problem, create the linkDir with permissions 0755.

Closes: https://github.com/containers/storage/issues/1270

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>