pkg/lockfile: wrap error in openLock
unix.Open returns a bare error (like EPERM). This is worked around in
on user, (*lockfile).lock, where we panic, but not in the other user,
createLockerForPath, and as a result such a bare error might be returned
from GetLockfile or GetROLockfile.

Wrap it into os.PathError and remove the workaround from lockfile.lock.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>