Fix Windows regression introduced by PR #1161
Before 50eb74a4b1156bd this code used unshare.IsRootless() which on Windows
always returns false (the behavior we want).

After 50eb74a4b1156bd, a condition was unintentionally inverted, allowing
Windows to function.

Commit 18803495e871 fixed the inversion, but unintentionally excluded Windows
since it used == 0 instead of <= 0 (Windows returns -1)

Move the logic behind a function with a comment since the Windows path is a
bit exotic.

In the future, the Windows path should likely be refactored to be more
intuitive; however, this will get things working for now.

Signed-off-by: Jason T. Greene <jason.greene@redhat.com>