drivers/overlay: Notify the user if the backing filesystem is not supported with OverlayFS
If the backing filesystem is not supported with OverlayFS as a container
storage driver/or graph driver then do not continue.
In this context, report the incompatibility by notifying the user of an
unsupported configuration. The filesystem type/or ID that is returned by
the statfs(2) system call is reported too, to assist further
troubleshooting efforts. It can be used to deduce the actual filesystem
used for the container storage graph directory.

Signed-off-by: Aaron Tomlin <atomlin@redhat.com>