Remove a manual layerStore.Touch() call from store.Shutdown()
It's completely unclear why this is necessary; the unmount calls
actually write to the separate mountpoints.json file, with a separate lock.

When this was originally introduced in c5b57afecd1ad63859525a96b290c2cfc59fd306 ,
there wasn't a separate mountpoints.json file, and store.go was manually
calling Touch() everywhere; so my best guess is that this is just an
artifact of making those two API changes, and the Touch() is not actually
necessary.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>