pkg/archive: Add functions for managing file flags
This adds platform-specific entry points to encode and decode file
flags. These are exported so that they can be used in buildah/copier. A
function which can be used to optionally reset the immutable flag on a
file or directory is also added and this will be used when applying
layer changes to trees with immutable files or directories.

These new entry points are implemented for FreeBSD and are noop stubs on
other platforms. The chflags approach could be used on darwin and other
BSD-derived platforms if needed.

On FreeBSD, file flag information is encoded for import/export using the
PAX SCHILY.fflags option. This is compatible with bsdtar (see
https://github.com/libarchive/libarchive). Reading the code for
libarchive, it also has support for encoding/decoding Linux file flags
so this seems like a good reference.

Signed-off-by: Doug Rabson <dfr@rabson.org>