fix hack/git-validation.sh script
It is impossible to have this passing because it did not run in CI and
invalid commit were merged since the fixed commit. The correct way
should be to always use the merge base commit instead, at least this
will only check commit that are actually part of the PR not all which is
wasteful. Also removed unused $PATH overwrite and $GITVALIDATE_FLAGS
var.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>